+++ src/index.js
@@ -123,4 +123,8 @@ export { default as Accordion } from './Accordion';
 export { default as SocialPreview } from './SocialPreview';
 export { default as GooglePreview } from './GooglePreview';
 export { Sidebar } from './Sidebar';
+export {
+  default as ListItemAction,
+  listItemActionBuilder,
+} from './ListItemAction';
 export { default as Subheader } from './Card/Subheader';