+++ testkit/puppeteer.js
@@ -210,6 +210,10 @@ import { sidebarDriverFactory } from '../src/Sidebar/Sidebar.uni.driver';
 
 export const sidebarTestkitFactory = puppeteerUniTestkitFactoryCreator(
   sidebarDriverFactory,
+import { ListItemActionDriverFactory } from '../src/ListItemAction/ListItemAction.uni.driver';
+
+export const ListItemActionTestkitFactory = puppeteerUniTestkitFactoryCreator(
+  ListItemActionDriverFactory,
 );
 
 import { cardSubheaderDriverFactory } from '../src/Card/Subheader/Subheader.uni.driver';